+++ sdk/storage/azfile/share/client.go
@@ -37,7 +37,8 @@ type Client base.Client[generated.ShareClient]
 // Note that the only share-level operations that support token credential authentication are CreatePermission and GetPermission.
 // Also note that ClientOptions.FileRequestIntent is currently required for token authentication.
 func NewClient(shareURL string, cred azcore.TokenCredential, options *ClientOptions) (*Client, error) {
-	authPolicy := runtime.NewBearerTokenPolicy(cred, []string{shared.TokenScope}, nil)
+	audience := base.GetAudience((*base.ClientOptions)(options))
+	authPolicy := runtime.NewBearerTokenPolicy(cred, []string{audience}, nil)
 	conOptions := shared.GetClientOptions(options)
 	plOpts := runtime.PipelineOptions{
 		PerRetry: []policy.Policy{authPolicy},