+++ sdk/storage/azdatalake/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 1.0.1 (Unreleased)
 
 ### Features Added
+* Encryption Scope For SAS
+* CPK for Datalake
+* Create SubDirectory Client
+* Service Version upgrade to 2021-06-08
 
 ### Breaking Changes
 
@@ -11,6 +15,7 @@
 * Block `SharedKeyCredential` authentication mode for non TLS protected endpoints. Fixes [#21841](https://github.com/Azure/azure-sdk-for-go/issues/21841).
 
 ### Other Changes
+* Updated version of azblob to 1.2.1
 
 * Updated azcore version to `1.9.1` and azidentity version to `1.4.0`.